// source/types/constants/PromoTargetInterest.ts
export enum PromoTargetInterest {
    BUSINESS_INDUSTRY = "Business & Industry",
    ENTERTAINMENT = "Entertainment",
    FITNESS_WELLNESS = "Fitness & Wellness",
    FOOD_DRINKS = "Food & Drinks",
    HOBBIES_ACTIVITIES = "Hobbies & Activities",
    SHOPPING_FASHION = "Shopping & Fashion",
    SPORT_OUTDOORS = "Sport & Outdoors",
    INTERNET = "Internet",
}